Coins gettable with bitfields.

+++ tilemap.h
@@ -40,8 +40,11 @@ public:
     void fill_coins();
     void draw_coins();
     void draw_coins(uint16_t offset_x);
-    bool get_coin(uint16_t x, uint8_t y);
+    bool get_coin(uint16_t x, uint16_t y);
+
     bool is_solid(uint16_t x, int8_t y);
+
+
